Fix minor bug
It seems like using the IMAGENET1K_V2 weights (via DEFAULT) leads to training not working. It took me forever to find this cuz I didn't expect V1 vs V2 to have such vast differences. The model or algo is really sensitive to this it seems which is weird cuz the algo works on ViTs thus proving to be generic enough. In the non-clean version of my code, I was simply using pretrained=True but since this has been deprecated, I switched to using ResNet50_Weights enum.